+++ cartridges/app_storefront_address_extend/cartridge/client/default/js/addressType.js
@@ -1,22 +1,6 @@
 
-const radioButtons = document.querySelectorAll('.js-form-radio-button');
-const businessAddress = document.querySelector('.js-business-address');
-const displayNone = 'd-none';
-const privateAddressString = 'Private Address';
-const businessAddressString = 'Business Address';
+var processInclude = require('base/util');
 
-radioButtons.forEach(radioButton => {
-    radioButton.onchange = () => {
-        
-        const addressType = radioButton.value;
-
-        switch (addressType) {
-            case privateAddressString:
-                businessAddress.classList.add(displayNone);
-                break;
-            case businessAddressString:
-                businessAddress.classList.remove(displayNone);
-                break;
-        }
-    }
+$(document).ready(function () {
+    processInclude(require('./checkout/addressType'));
 });
